refactor: Session에서 maxUserEnrollment 상태 NextStepUsers로 책임 분리
