+++ src/main/java/nextstep/courses/domain/Session.java
@@ -10,8 +10,7 @@ public class Session {
   private final Long id;
   private final SessionPayment sessionPayment;
   private SessionStatus sessionStatus;
-  private final MaxUserEnrollment maxUserEnrollment;
-  private final NextStepUsers nextStepUsers = new NextStepUsers();
+  private final NextStepUsers nextStepUsers;
   private final SessionDate startDate;
   private final SessionDate closedDate;
   private final SessionCoverUrl sessionCoverUrl;
@@ -23,7 +22,7 @@ public class Session {
     this.id = id;
     this.sessionPayment = sessionPayment;
     this.sessionStatus = sessionStatus;
-    this.maxUserEnrollment = new MaxUserEnrollment(maxUserEnrollment);
+    this.nextStepUsers = new NextStepUsers(maxUserEnrollment);
     this.closedDate = new SessionDate(closedDate);
     this.sessionCoverUrl = new SessionCoverUrl(sessionCoverUrl);
   }
@@ -38,7 +37,6 @@ public class Session {
 
   public void processEnrollment(NextStepUser nextStepUser) {
     sessionStatus.validateAcceptingStatus();
-    maxUserEnrollment.validateUserEnrollment(nextStepUsers);
 
     nextStepUsers.enroll(nextStepUser);
   }
